tool-src/requirements.txt
Pillow>=9.4.0
opencv-python>=4.7.0
numpy>=1.24.0
scikit-image>=0.19.3
scikit-learn>=1.2.0
scipy>=1.10.0
tqdm>=4.64.1
